+++ xyui/avalonia/src/XYUI.Avalonia/Controls/XYCheckbox.cs
@@ -1,8 +1,0 @@
-using Avalonia.Controls;
-
-namespace XYUI.Avalonia.Controls;
-
-public class XYCheckbox : CheckBox
-{
-    public XYCheckbox() => Classes.Add("xyui-checkbox");
-}