+++ xyui/avalonia/src/XYUI.Avalonia/Controls/XYUI2/XYUI2-06-Checkbox/XYCheckbox.cs
@@ -1,0 +1,12 @@
+using Avalonia.Controls;
+
+namespace XYUI.Avalonia.Controls;
+
+public partial class XYCheckbox : CheckBox
+{
+    public XYCheckbox()
+    {
+        Classes.Add("xyui-checkbox");
+        Focusable = true;
+    }
+}